requirements.txt
streamlit
Pillow
pdf2image